Fix Gen 8 temperature reading (#851)
* Fix type casting in WMI

* Fix type casting in WMI

* Sometimes, everyone is stupid

* Fix gen 8 legion zone version reading

+++ LenovoLegionToolkit.Lib/Controllers/DisplayBrightnessController.cs
@@ -8,5 +8,5 @@ public class DisplayBrightnessController
     public Task SetBrightnessAsync(int brightness) => WMI.CallAsync(@"root\WMI",
         $"SELECT * FROM WmiMonitorBrightnessMethods",
         "WmiSetBrightness",
-        new() { { "Timeout", (uint)1 }, { "Brightness", (byte)brightness } });
+        new() { { "Timeout", 1 }, { "Brightness", brightness } });
 }
